feat: add Multiple Negatives Ranking Loss.

+++ source/metric/MRRMetric.py
@@ -16,8 +16,8 @@ class MRRMetric(Metric):
         :param x2: a tensors with shape [batch_size, hidden_size].
         :return: the cosine similarity matrix with shape [batch_size, batch_size].
         """
-        x1 = x1 / torch.norm(x1, dim=1, p=2, keepdim=True, dtype=torch.float32)
-        x2 = x2 / torch.norm(x2, dim=1, p=2, keepdim=True, dtype=torch.float32)
+        x1 = x1 / torch.norm(x1, dim=1, p=2, keepdim=True)
+        x2 = x2 / torch.norm(x2, dim=1, p=2, keepdim=True)
         return torch.matmul(x1, x2.t())
 
     def update(self, r1, r2):
